pyproject.toml
[project]
name = "pr-agent-review-bot"
version = "0.1.0"
description = "Multi-agent PR reviewer bot for GitHub using FastAPI, webhook handling, and optional LLM orchestration."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name="Devang Banda"}]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "PyJWT>=2.8",
  "tenacity>=8.2",
  "python-dotenv>=1.0",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
  "mypy>=1.10",
  "types-requests>=2.32.0.20240712",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
